requirements.txt
fastapi
uvicorn
black
supabase
load_dotenv